Update and rename sql.go to mysql.go
